+++ androidapp/Vt/app/src/main/java/com/vt/vt/core/data/source/remote/income/IncomeRequest.kt
@@ -1,6 +1,7 @@
 package com.vt.vt.core.data.source.remote.income
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class IncomeRequest(
 
@@ -8,7 +9,7 @@ data class IncomeRequest(
 
     @field:SerializedName("income_category_id") val incomeCategoryId: Int? = null,
 
-    @field:SerializedName("amount") val amount: Double,
+    @field:SerializedName("amount") val amount: BigDecimal,
 
     @field:SerializedName("remarks") val remarks: String? = null
 )
